move selected output functions to and external class in externalio
